+++ maggy/core/environment/__init__.py
@@ -1,20 +1,15 @@
-import os
-
-from maggy.core.environment import abstractenvironment, baseenvironment
-
-AbstractEnvironment = abstractenvironment.AbstractEnvironment
-# check which environment to use
-if "REST_ENDPOINT" in os.environ:
-    print("You are running maggy on Hopsworks.")
-
-    from maggy.core.environment import hopsenvironment
-
-    HopsEnvironment = hopsenvironment.HopsEnvironment
-
-    __all__ = ["AbstractEnvironment", "HopsEnvironment"]
-
-else:
-    print("You are running maggy without hopsworks.")
-    BaseEnvironment = baseenvironment.BaseEnvironment
-
-    __all__ = ["AbstractEnvironment", "BaseEnvironment"]
+#
+#   Copyright 2021 Logical Clocks AB
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
