+++ maggy/optimizer/abstractoptimizer.py
@@ -19,7 +19,7 @@ import time
 from abc import ABC, abstractmethod
 from datetime import datetime
 
-from maggy.core.environment_singleton import environment_singleton
+from maggy.core.environment.singleton import EnvSing
 from maggy.pruner import Hyperband
 from maggy.trial import Trial
 
@@ -49,8 +49,6 @@ class AbstractOptimizer(ABC):
 
         # helper variable to calculate time needed for calculating next suggestion
         self.sampling_time_start = 0.0
-
-        self.env = environment_singleton()
 
     @abstractmethod
     def initialize(self):
@@ -88,12 +86,12 @@ class AbstractOptimizer(ABC):
         :param exp_dir: path of experiment directory
         :rtype exp_dir: str
         """
-
+        env = EnvSing.get_instance()
         # configure logger
         self.log_file = exp_dir + "/optimizer.log"
-        if not self.env.exists(self.log_file):
-            self.env.dump("", self.log_file)
-        self.fd = self.env.open_file(self.log_file, flags="w")
+        if not env.exists(self.log_file):
+            env.dump("", self.log_file)
+        self.fd = env.open_file(self.log_file, flags="w")
         self._log("Initialized Optimizer Logger")
 
     def _initialize(self, exp_dir):
@@ -127,7 +125,7 @@ class AbstractOptimizer(ABC):
     def _log(self, msg):
         if self.fd and not self.fd.closed:
             msg = datetime.now().isoformat() + ": " + str(msg)
-            self.fd.write(environment_singleton().str_or_byte(msg + "\n"))
+            self.fd.write(EnvSing.get_instance().str_or_byte(msg + "\n"))
 
     def _close_log(self):
         if not self.fd.closed:
